# docs/openapi.yaml
openapi: 3.0.0
servers:
  - description: CERCLES API
    url: http://localhost:8080
info:
  version: "1.0.0"
  title: CERCLES API Documentation
  description: >-
    Esta API proporciona los endpoints para la aplicación CERCLES.
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Ingrese su token JWT para autenticar las solicitudes"
security:
  - BearerToken: []
...
paths:
...
  /api/github/validar-organizacion:
    post:
      tags:
        - GitHub
      summary: Validar organización en GitHub
      description: >
        Este endpoint permite validar si una organización de GitHub cumple con los requisitos específicos 
        como tener ciertos miembros, un profesor configurado, y si la cuenta 'professorat-amep' es miembro y admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                organizacionUrl:
                  type: string
                  description: URL de la organización en GitHub.
                  example: "https://github.com/orgName"
                miembrosIds:
                  type: array
                  items:
                    type: integer
                  description: Lista de IDs de los miembros que deben ser validados.
                profesorId:
                  type: integer
                  description: ID del profesor que será validado en la organización.
      responses:
        "200":
          description: Validación exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  todosUsuariosGitConfigurados:
                    type: boolean
                    description: Indica si todos los usuarios tienen configurado su GitHub.
                  profesorGitConfigurado:
                    type: boolean
                    description: Indica si el profesor tiene configurado su GitHub.
                  professoratEsMiembro:
                    type: boolean
                    description: Indica si la cuenta 'professorat-amep' es miembro de la organización.
                  professoratEsAdmin:
                    type: boolean
                    description: Indica si la cuenta 'professorat-amep' es admin de la organización.
                  todosMiembrosEnOrganizacion:
                    type: boolean
                    description: Indica si todos los miembros especificados están en la organización.
                  profesorEnOrganizacion:
                    type: boolean
                    description: Indica si el profesor está en la organización.
        "403":
          description: Usuario no autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/github/confirmar-organizacion:
    post:
      tags:
        - GitHub
      summary: Confirmar organización en GitHub
      description: >
        Este endpoint asigna una organización a un equipo después de haber sido validada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                equipoId:
                  type: integer
                  description: ID del equipo al cual se asociará la organización.
                organizacionUrl:
                  type: string
                  description: URL de la organización en GitHub.
      responses:
        "200":
          description: Organización asignada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /api/github/callback:
    post:
      tags:
        - GitHub
      summary: Conectar usuario con GitHub
      description: >
        Este endpoint gestiona el callback de GitHub OAuth para conectar la cuenta de GitHub del usuario con su perfil en el sistema.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: Código de autorización proporcionado por GitHub.
                  example: "123abc456def"
      responses:
        "200":
          description: Conexión exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensaje de confirmación.
                    example: "Cuenta de GitHub asociada exitosamente"
                  githubUsername:
                    type: string
                    description: Nombre de usuario en GitHub.
                    example: "johnDoeGitHub"
        "400":
          description: Código faltante o inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Código de GitHub faltante"
        "403":
          description: Usuario no autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Usuario no autenticado"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error interno del servidor"
  /api/github/disconnect:
    delete:
      tags:
        - GitHub
      summary: Desconectar cuenta de GitHub
      description: >
        Este endpoint permite a un usuario desconectar su cuenta de GitHub del sistema.
      responses:
        "200":
          description: Cuenta desconectada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "403":
          description: Usuario no autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/github/user-data:
    get:
      tags:
        - GitHub
      summary: Obtener datos del usuario en GitHub
      description: >
        Este endpoint permite a un usuario autenticado recuperar información sobre sus repositorios y organizaciones asociadas en GitHub.
      responses:
        "200":
          description: Datos obtenidos exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  repositorios:
                    type: array
                    description: Lista de repositorios asociados al usuario.
                    items:
                      type: object
                      additionalProperties: true
                  organizaciones:
                    type: array
                    description: Lista de organizaciones a las que pertenece el usuario.
                    items:
                      type: object
                      additionalProperties: true
        "403":
          description: Usuario no autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Usuario no autenticado"
        "404":
          description: Datos no encontrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Datos no encontrados para este usuario."
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error interno del servidor"